# Simplify the code a bit using async/await