--- a/.eslintrc
+++ b/.eslintrc
@@ -7,6 +7,7 @@
   "extends": ["airbnb-base"],
   "rules": {
     "arrow-parens": 0,
-    "eslint-disable-next": 0
+    "eslint-disable-next": 0,
+    "no-restricted-syntax": 0
   }
 }