--- a/lib/index.js
+++ b/lib/index.js
@@ -15,9 +15,16 @@ exports.validate = (schema = {}, options = {}, joi = {}) => {
     const evOptions = mergeEvOptions(options);
     const joiOptions = mergeJoiOptions(joi, options.context, request);
 
-    const validate = (parameter) => schema[parameter].validateAsync(request[parameter], joiOptions)
-      .then(result => handleMutation(request, parameter, result.value, evOptions.context))
-      .catch(error => ({ [parameter]: error.details }));
+    async function validate(parameter) {
+      let result;
+      try {
+        result = await schema[parameter].validateAsync(request[parameter], joiOptions);
+      } catch (error) {
+        return { [parameter]: error.details };
+      }
+      handleMutation(request, parameter, result.value, evOptions.context);
+      return undefined;
+    }
 
     const hasErrors = (errors) => (errors ? new ValidationError(errors, evOptions) : null);