--- a/lib/index.js
+++ b/lib/index.js
@@ -26,19 +26,25 @@ exports.validate = (schema = {}, options = {}, joi = {}) => {
     return undefined;
   }
 
-  return (request, response, next) => {
-    const joiOptions = options.context ? { ...baseJoiOptions, context: request } : baseJoiOptions;
+  const hasErrors = (errors) => (errors ? new ValidationError(errors, evOptions) : null);
 
-    const hasErrors = (errors) => (errors ? new ValidationError(errors, evOptions) : null);
+  return async (request, response, next) => {
+    try {
+      const joiOptions = options.context ? { ...baseJoiOptions, context: request } : baseJoiOptions;
 
-    const promises = parameters
-      .filter(parameter => request[parameter] && schema[parameter])
-      .map(parameter => validate(request, parameter, joiOptions));
+      const promises = parameters
+        .filter(parameter => request[parameter] && schema[parameter])
+        .map(parameter => validate(request, parameter, joiOptions));
 
-    return Promise.all(promises)
-      .then(e => clean(e))
-      .then(e => keyByField(e, options.keyByField))
-      .then(r => next(hasErrors(r)));
+      let e = await Promise.all(promises);
+      e = clean(e);
+      const r = keyByField(e, options.keyByField);
+
+      next(hasErrors(r));
+    } catch (err) {
+      // Make sure that we don't cause an unhandled rejection if there's an error in the code:
+      next(err);
+    }
   };
 };