--- a/lib/index.js
+++ b/lib/index.js
@@ -26,8 +26,6 @@ exports.validate = (schema = {}, options = {}, joi = {}) => {
     return undefined;
   }
 
-  const hasErrors = (errors) => (errors ? new ValidationError(errors, evOptions) : null);
-
   return async (request, response, next) => {
     try {
       const joiOptions = options.context ? { ...baseJoiOptions, context: request } : baseJoiOptions;
@@ -36,11 +34,13 @@ exports.validate = (schema = {}, options = {}, joi = {}) => {
         .filter(parameter => request[parameter] && schema[parameter])
         .map(parameter => validate(request, parameter, joiOptions));
 
-      let e = await Promise.all(promises);
-      e = clean(e);
-      const r = keyByField(e, options.keyByField);
+      const errors = keyByField(clean(await Promise.all(promises)), options.keyByField);
 
-      next(hasErrors(r));
+      if (errors) {
+        next(new ValidationError(errors, evOptions));
+      } else {
+        next();
+      }
     } catch (err) {
       // Make sure that we don't cause an unhandled rejection if there's an error in the code:
       next(err);