--- a/lib/index.js
+++ b/lib/index.js
@@ -16,14 +16,13 @@ exports.validate = (schema = {}, options = {}, joi = {}) => {
   const baseJoiOptions = { ...joi, warnings: true };
 
   async function validate(request, parameter, joiOptions) {
-    let result;
     try {
-      result = await schema[parameter].validateAsync(request[parameter], joiOptions);
+      const result = await schema[parameter].validateAsync(request[parameter], joiOptions);
+      handleMutation(request, parameter, result.value, evOptions.context);
+      return undefined;
     } catch (error) {
       return { [parameter]: error.details };
     }
-    handleMutation(request, parameter, result.value, evOptions.context);
-    return undefined;
   }
 
   return async (request, response, next) => {