--- a/lib/index.js
+++ b/lib/index.js
@@ -17,8 +17,8 @@ exports.validate = (schema = {}, options = {}, joi = {}) => {
 
   async function validate(request, parameter, joiOptions) {
     try {
-      const result = await schema[parameter].validateAsync(request[parameter], joiOptions);
-      handleMutation(request, parameter, result.value, evOptions.context);
+      const { value } = await schema[parameter].validateAsync(request[parameter], joiOptions);
+      handleMutation(request, parameter, value, evOptions.context);
       return undefined;
     } catch (error) {
       return { [parameter]: error.details };