--- a/lib/index.js
+++ b/lib/index.js
@@ -29,9 +29,13 @@ exports.validate = (schema = {}, options = {}, joi = {}) => {
     try {
       const joiOptions = options.context ? { ...baseJoiOptions, context: request } : baseJoiOptions;
 
-      const promises = parameters
-        .filter(parameter => request[parameter] && schema[parameter])
-        .map(parameter => validate(request, parameter, joiOptions));
+      const promises = [];
+      // eslint-disable-next-line no-restricted-syntax
+      for (const parameter of parameters) {
+        if (request[parameter] && schema[parameter]) {
+          promises.push(validate(request, parameter, joiOptions));
+        }
+      }
 
       const errors = keyByField(clean(await Promise.all(promises)), options.keyByField);