--- a/lib/index.js
+++ b/lib/index.js
@@ -30,7 +30,6 @@ exports.validate = (schema = {}, options = {}, joi = {}) => {
       const joiOptions = options.context ? { ...baseJoiOptions, context: request } : baseJoiOptions;
 
       const promises = [];
-      // eslint-disable-next-line no-restricted-syntax
       for (const parameter of parameters) {
         if (request[parameter] && schema[parameter]) {
           promises.push(validate(request, parameter, joiOptions));