--- a/lib/joi.js
+++ b/lib/joi.js
@@ -6,10 +6,4 @@ const schema = parameters
 
 const joiSchema = Joi.object(schema).required().min(1);
 
-const mergeJoiOptions = (joi, ctx = null, request) => {
-  const context = ctx ? request : null;
-  return { ...joi, warnings: true, context };
-};
-
 exports.joiSchema = joiSchema;
-exports.mergeJoiOptions = mergeJoiOptions;