… __tests__/unit/joiOptions.test.js