I've had a bit of trouble wrapping my head around exactly what's going on in `index.js`. Adopting `async`/`await` seems to help a bit. Added some drive-by optimizations as well so more stuff is precomputed outside the middleware itself.

This makes the middleware function `async`, but it doesn't ever return a rejected promise because its body is wrapped in a try/catch. This should still be compatible with both vanilla `express` and `express-promise-router`.

Drive-by fixes https://github.com/AndrewKeig/express-validation/issues/153

@AndrewKeig, this is a bit of an invasive refactoring, so I would appreciate a sign-off from you :grimacing: :crossed_fingers: