@AndrewKeig 2022-04-09
Hey, I’m cool with it as long as no impact on performance, see the load tests to confirm any issues