@AndrewKeig 2022-04-09
Also, maybe bump the major version as this might break old versions of node, thanks for supporting this by the way, much appreciated