@papandreou 2022-04-09
Tried running the load tests with node.js 16.14.2 before and after applying this change:

### Before

```
[express-validation]$ ./load/post.sh 
Running 10s test @ http://localhost:3000/login
10 connections


┌─────────┬──────┬──────┬───────┬──────┬─────────┬─────────┬───────┐
│ Stat    │ 2.5% │ 50%  │ 97.5% │ 99%  │ Avg     │ Stdev   │ Max   │
├─────────┼──────┼──────┼───────┼──────┼─────────┼─────────┼───────┤
│ Latency │ 2 ms │ 2 ms │ 5 ms  │ 7 ms │ 2.43 ms │ 1.06 ms │ 28 ms │
└─────────┴──────┴──────┴───────┴──────┴─────────┴─────────┴───────┘
┌───────────┬────────┬────────┬─────────┬─────────┬─────────┬────────┬────────┐
│ Stat      │ 1%     │ 2.5%   │ 50%     │ 97.5%   │ Avg     │ Stdev  │ Min    │
├───────────┼────────┼────────┼─────────┼─────────┼─────────┼────────┼────────┤
│ Req/Sec   │ 1794   │ 1794   │ 3625    │ 3721    │ 3363.82 │ 561.74 │ 1794   │
├───────────┼────────┼────────┼─────────┼─────────┼─────────┼────────┼────────┤
│ Bytes/Sec │ 858 kB │ 858 kB │ 1.73 MB │ 1.78 MB │ 1.61 MB │ 269 kB │ 858 kB │
└───────────┴────────┴────────┴─────────┴─────────┴─────────┴────────┴────────┘

Req/Bytes counts sampled once per second.
# of samples: 11

0 2xx responses, 37000 non 2xx responses
37k requests in 11.02s, 17.7 MB read
```

### After

```
[express-validation]$ ./load/post.sh 
Running 10s test @ http://localhost:3000/login
10 connections


┌─────────┬──────┬──────┬───────┬──────┬────────┬─────────┬───────┐
│ Stat    │ 2.5% │ 50%  │ 97.5% │ 99%  │ Avg    │ Stdev   │ Max   │
├─────────┼──────┼──────┼───────┼──────┼────────┼─────────┼───────┤
│ Latency │ 2 ms │ 2 ms │ 5 ms  │ 7 ms │ 2.4 ms │ 1.05 ms │ 26 ms │
└─────────┴──────┴──────┴───────┴──────┴────────┴─────────┴───────┘
┌───────────┬────────┬────────┬─────────┬─────────┬─────────┬────────┬────────┐
│ Stat      │ 1%     │ 2.5%   │ 50%     │ 97.5%   │ Avg     │ Stdev  │ Min    │
├───────────┼────────┼────────┼─────────┼─────────┼─────────┼────────┼────────┤
│ Req/Sec   │ 1804   │ 1804   │ 3667    │ 3833    │ 3420.9  │ 623.18 │ 1804   │
├───────────┼────────┼────────┼─────────┼─────────┼─────────┼────────┼────────┤
│ Bytes/Sec │ 863 kB │ 863 kB │ 1.75 MB │ 1.83 MB │ 1.63 MB │ 298 kB │ 862 kB │
└───────────┴────────┴────────┴─────────┴─────────┴─────────┴────────┴────────┘

Req/Bytes counts sampled once per second.
# of samples: 10

0 2xx responses, 34206 non 2xx responses
34k requests in 10.02s, 16.4 MB read
```